No need to check if term exists

+++ includes/member-functions.php
@@ -475,12 +475,19 @@ function rcp_print_status( $user_id = 0, $echo = true  ) {
 function rcp_set_status( $user_id, $new_status ) {
 
 	$old_status = get_user_meta( $user_id, 'rcp_status', true );
+
 	if( ! $old_status ) {
 		$old_status = __( 'Free', 'rcp' );
 	}
 
+	if( $old_status == $new_status ) {
+		return false;
+	}
+
 	if( update_user_meta( $user_id, 'rcp_status', $new_status ) ) {
+
 		delete_user_meta( $user_id, '_rcp_expired_email_sent');
+
 		do_action( 'rcp_set_status', $new_status, $user_id );
 
 		// Record the status change
